move support-btn component in common folder
