+++ .gitignore
@@ -1,7 +1,6 @@
 node_modules/
 build/
 bower_components/
-dist/
 .tmp
 .idea
 *.sublime-project
