+++ src_ts/components/app-shell/header/page-header.ts
@@ -4,7 +4,7 @@ import '@polymer/polymer/lib/elements/dom-if.js';
 import '@polymer/iron-flex-layout/iron-flex-layout.js';
 import '@polymer/app-layout/app-toolbar/app-toolbar.js';
 import '@polymer/paper-icon-button/paper-icon-button.js';
-import './support-btn';
+import '../../common/support-btn';
 
 import { connect } from 'pwa-helpers/connect-mixin.js';
 import {store, RootState} from "../../../store";
@@ -93,7 +93,7 @@ class PageHeader extends connect(store)(GestureEventListeners(PolymerElement)) {
           <!--<countries-dropdown id="countries" countries="[[countries]]"-->
                               <!--current-country="[[profile.country]]"></countries-dropdown>-->
 
-          <!-- <support-btn></support-btn> -->
+          <support-btn></support-btn> 
 
           <!--<etools-profile-dropdown-->
               <!--sections="[[allSections]]"-->
